refactor: remove mui, replace theme structure and icons
