+++ dashboard/package.json
@@ -6,8 +6,6 @@
   "dependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
-    "@mui/icons-material": "^7.1.1",
-    "@mui/material": "^7.0.1",
     "@tailwindcss/vite": "^4.1.3",
     "@types/d3": "^7.4.3",
     "@types/node": "^20.0.0",
